# src/bounded-contexts/skills-catalog/tech-skills/controllers/tech-skill.controller.ts
/**
 * Tech Skill Controller
 * Endpoints for tech skills
 */

import { Controller, Get, Query, Param } from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse } from '@nestjs/swagger';
import { SdkExport } from '@/bounded-contexts/platform/common/decorators/sdk-export.decorator';
import { SkillQueryService } from '../services/skill-query.service';
import { SkillSearchService } from '../services/skill-search.service';
import { TechSkillsQueryService } from '../services/tech-skills-query.service';
import { Public } from '@/bounded-contexts/identity/auth/decorators/public.decorator';
import type { SkillType } from '../interfaces';
import type { TechSkill } from '../dtos';
import { TechSkillDto } from '@/shared-kernel';

@SdkExport({
  tag: 'tech-skills',
  description: 'Tech Skills API',
  requiresAuth: false,
})
@ApiTags('tech-skills')
@Controller('v1/tech-skills')
export class TechSkillController {
  constructor(
    private readonly skillQuery: SkillQueryService,
    private readonly skillSearch: SkillSearchService,
    private readonly queryService: TechSkillsQueryService,
  ) {}

  /** Get all skills */
  @Get()
  @Public()
  @ApiOperation({ summary: 'Get all tech skills' })
  @ApiResponse({ status: 200, type: [TechSkillDto] })
  @ApiResponse({ status: 200, description: 'List of tech skills' })
  async getSkills(): Promise<TechSkill[]> {
    return this.skillQuery.getAllSkills();
  }

  /** Search skills */
  @Get('search')
  @Public()
  @ApiOperation({ summary: 'Search tech skills' })
  @ApiResponse({ status: 200, type: [TechSkillDto] })
  @ApiResponse({ status: 200, description: 'Search results' })
  async searchSkills(
    @Query('q') query: string,
    @Query('limit') limit?: string,
  ): Promise<TechSkill[]> {
    return this.skillSearch.searchSkills(
      query,
      limit ? parseInt(limit, 10) : 20,
    );
  }

  /** Get skills by type */
  @Get('type/:type')
  @Public()
  @ApiOperation({ summary: 'Get skills by type' })
  @ApiResponse({ status: 200, type: [TechSkillDto] })
  @ApiResponse({ status: 200, description: 'List of skills by type' })
  async getSkillsByType(
    @Param('type') type: SkillType,
    @Query('limit') limit?: string,
  ): Promise<TechSkill[]> {
    return this.queryService.getSkillsByType(
      type,
      limit ? parseInt(limit, 10) : 50,
    );
  }
}