# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ProFile API",
    "description": "ProFile Resume & Portfolio Management API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    }
  ],
...
  "paths": {
    "/api/v1/resumes": {
      "get": {
        "operationId": "resumes_getAllUserResumes",
        "tags": [
          "resumes"
        ],
        "summary": "Get all resumes for current user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get all resumes for current user"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "post": {
        "operationId": "resumes_createResumeForUser",
        "tags": [
          "resumes"
        ],
        "summary": "Create a new resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Create a new resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/resumes/slots": {
      "get": {
        "operationId": "resumes_getRemainingSlots",
        "tags": [
          "resumes"
        ],
        "summary": "Get remaining resume slots for current user",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Get remaining resume slots for current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeSlotsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/resumes/{id}/full": {
      "get": {
        "operationId": "resumes_getResumeByIdWithAllSections",
        "tags": [
          "resumes"
        ],
        "summary": "Get a resume with all sections",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a resume with all sections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeFullResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/resumes/{id}": {
      "get": {
        "operationId": "resumes_getResumeByIdForUser",
        "tags": [
          "resumes"
        ],
        "summary": "Get a specific resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a specific resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeFullResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "patch": {
        "operationId": "resumes_updateResumeForUser",
        "tags": [
          "resumes"
        ],
        "summary": "Update a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update a resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "delete": {
        "operationId": "resumes_deleteResumeForUser",
        "tags": [
          "resumes"
        ],
        "summary": "Delete a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/resumes/{resumeId}/skills": {
      "get": {
        "operationId": "resumes_findAllSkillsForResume",
        "tags": [
          "resumes"
        ],
        "summary": "Get all skills for a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all skills for a resume"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
    "/api/v1/resumes/{resumeId}/skills/bulk": {
      "post": {
        "operationId": "resumes_createBulk",
        "tags": [
          "resumes"
        ],
        "summary": "Create multiple skills at once",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create multiple skills at once"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
    "/api/v1/tech-skills/search": {
      "get": {
        "operationId": "tech-skills_searchSkills",
        "tags": [
          "tech-skills"
        ],
        "summary": "Search tech skills",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search tech skills"
          }
        }
      }
    },
    "/api/v1/tech-skills/type/{type}": {
      "get": {
        "operationId": "tech-skills_getSkillsByType",
        "tags": [
          "tech-skills"
        ],
        "summary": "Get skills by type",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get skills by type"
          }
        }
      }
    },
    "/api/v1/skills/manage/resume/{resumeId}": {
      "get": {
        "operationId": "skills_listSkillsForResume",
        "tags": [
          "skills"
        ],
        "summary": "List all skills for a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all skills for a resume"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      },
      "post": {
        "operationId": "skills_addSkillToResume",
        "tags": [
          "skills"
        ],
        "summary": "Add a skill to a resume",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Add a skill to a resume",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeSkillResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "DeleteResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
      "ResumeFullResponseDto": {
        "type": "object",
        "properties": {
          "experiences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceResponseDto"
            }
          },
          "educations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EducationResponseDto"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillResponseDto"
            }
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        },
        "required": [
          "experiences",
          "educations",
          "skills"
        ]
      },
...
      "ResumeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "title": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "example": "false"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "language": {
            "type": "string"
          },
          "targetRole": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "isPublic",
          "createdAt",
          "updatedAt"
        ]
      },
      "ResumeSlotsResponseDto": {
        "type": "object",
        "properties": {
          "used": {
            "type": "number",
            "example": "2"
          },
          "limit": {
            "type": "number",
            "example": "4"
          },
          "remaining": {
            "type": "number",
            "example": "2"
          }
        },
        "required": [
          "used",
          "limit",
          "remaining"
        ]
      },
      "ResumeSkillResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "resumeId": {
            "type": "string",
            "example": "clxxx..."
          },
          "name": {
            "type": "string",
            "example": "TypeScript"
          },
          "order": {
            "type": "number",
            "example": "0"
          },
          "level": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "resumeId",
          "name",
          "order"
        ]
      },
...
      "ExperienceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "order": {
            "type": "number",
            "example": "0"
          },
          "location": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "current": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "order"
        ]
      },
      "EducationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "institution": {
            "type": "string"
          },
          "order": {
            "type": "number",
            "example": "0"
          },
          "degree": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "gpa": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "institution",
          "order"
        ]
      },
      "SkillResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clxxx..."
          },
          "name": {
            "type": "string",
            "example": "TypeScript"
          },
          "order": {
            "type": "number",
            "example": "0"
          },
          "level": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "order"
        ]
      },
...
    }
  }
}